Feat: warmup service connect
